Rename Lesson and Error in Content
