+++ src/lib/layout/index.ts
@@ -1,7 +1,7 @@
 export { default as Header } from "./_layout/Header.svelte";
 export { default as Content } from "./_layout/Content.svelte";
-export { default as LessonCard } from "./_layout/LessonCard.svelte";
-export { default as ErrorCard } from "./_layout/ErrorCard.svelte";
+export { default as Lesson } from "./_layout/Lesson.svelte";
+export { default as Error } from "./_layout/Error.svelte";
 export { default as AddDialog } from "./_layout/AddDialog.svelte";
 export { default as InfoDialog } from "./_layout/InfoDialog.svelte";
 export { default as WeekPicker } from "./_layout/WeekPicker.svelte";
